+++ src/lint/mod.rs
@@ -7,24 +7,33 @@ use std::path::{Path, PathBuf};
 use std::str::FromStr;
 use lazy_static::lazy_static;
 use log::{debug, error, trace};
+use rules::linelength::{BreakBeforeBinaryOpOptions,
+    BreakFuncCallOpenParenOptions,
+    BreakMethodOutputOptions,
+    BreakConditionalExpressionOptions};
 use serde::{Deserialize, Serialize};
 use regex::Regex;
 use rules::{instantiate_rules, CurrentRules, RuleType};
 use rules::{spacing::{SpReservedOptions,
-                      SpBraceOptions,
-                      SpPunctOptions,
-                      SpBinopOptions,
-                      NspFunparOptions,
-                      SpTernaryOptions,
-                      SpPtrDeclOptions,
-                      NspPtrDeclOptions,
-                      NspInparenOptions,
-                      NspUnaryOptions,
-                      NspTrailingOptions},
-            indentation::{LongLineOptions, IndentSizeOptions, IndentCodeBlockOptions,
-                          IndentNoTabOptions, IndentClosingBraceOptions, IndentParenExprOptions,
-                          IndentSwitchCaseOptions, IndentEmptyLoopOptions},
-                    };
+                SpBraceOptions,
+                SpPunctOptions,
+                SpBinopOptions,
+                NspFunparOptions,
+                SpTernaryOptions,
+                SpPtrDeclOptions,
+                NspPtrDeclOptions,
+                NspInparenOptions,
+                NspUnaryOptions,
+                NspTrailingOptions},
+            indentation::{LongLineOptions,
+                IndentSizeOptions,
+                IndentCodeBlockOptions,
+                IndentNoTabOptions,
+                IndentClosingBraceOptions,
+                IndentParenExprOptions,
+                IndentSwitchCaseOptions,
+                IndentEmptyLoopOptions,
+                IndentContinuationLineOptions}};
 use crate::analysis::{DMLError, IsolatedAnalysis, LocalDMLError, ZeroRange};
 use crate::analysis::parsing::tree::TreeElement;
 use crate::file_management::CanonPath;
@@ -109,6 +118,16 @@ pub struct LintCfg {
     pub indent_switch_case: Option<IndentSwitchCaseOptions>,
     #[serde(default)]
     pub indent_empty_loop: Option<IndentEmptyLoopOptions>,
+    #[serde(default)]
+    pub indent_continuation_line: Option<IndentContinuationLineOptions>,
+    #[serde(default)]
+    pub break_func_call_open_paren: Option<BreakFuncCallOpenParenOptions>,
+    #[serde(default)]
+    pub break_conditional_expression: Option<BreakConditionalExpressionOptions>,
+    #[serde(default)]
+    pub break_method_output: Option<BreakMethodOutputOptions>,
+    #[serde(default)]
+    pub break_before_binary_op: Option<BreakBeforeBinaryOpOptions>,
     #[serde(default = "get_true")]
     pub annotate_lints: bool,
 }
@@ -154,6 +173,11 @@ impl Default for LintCfg {
             indent_paren_expr: Some(IndentParenExprOptions{}),
             indent_switch_case: Some(IndentSwitchCaseOptions{indentation_spaces: INDENTATION_LEVEL_DEFAULT}),
             indent_empty_loop: Some(IndentEmptyLoopOptions{indentation_spaces: INDENTATION_LEVEL_DEFAULT}),
+            indent_continuation_line: Some(IndentContinuationLineOptions{indentation_spaces: INDENTATION_LEVEL_DEFAULT}),
+            break_func_call_open_paren: Some(BreakFuncCallOpenParenOptions{indentation_spaces: INDENTATION_LEVEL_DEFAULT}),
+            break_method_output: Some(BreakMethodOutputOptions{}),
+            break_conditional_expression: Some(BreakConditionalExpressionOptions{}),
+            break_before_binary_op: Some(BreakBeforeBinaryOpOptions{}),
             annotate_lints: true,
         }
     }
@@ -218,7 +242,7 @@ impl LinterAnalysis {
     }
 }
 
-pub fn begin_style_check(ast: TopAst, file: &str, rules: &CurrentRules) -> Result<Vec<DMLStyleError>, Error> {
+fn begin_style_check(ast: TopAst, file: &str, rules: &CurrentRules) -> Result<Vec<DMLStyleError>, Error> {
     let (mut invalid_lint_annot, lint_annot) = obtain_lint_annotations(file);
     let mut linting_errors: Vec<DMLStyleError> = vec![];
     ast.style_check(&mut linting_errors, rules, AuxParams { depth: 0 });
@@ -290,8 +314,7 @@ fn obtain_lint_annotations(file: &str) -> (Vec<DMLStyleError>,
         last_line = row;
         if let Some(capture) = LINT_ANNOTATION.captures(line) {
             let has_pre = capture.get(1)
-                .map_or(false,
-                        |m|!m.is_empty() &&
+                .is_some_and(|m|!m.is_empty() &&
                         JUST_WHITESPACE.captures(m.as_str()).is_none());
             let op_capture = capture.get(2).unwrap();
             let operation = match op_capture.as_str() {
@@ -397,7 +420,7 @@ fn remove_disabled_lints(errors: &mut Vec<DMLStyleError>,
             !annotations.whole_file.contains(
                 &LintAnnotation::Allow(error.rule_type)) &&
                 !annotations.line_specific.get(&error.error.range.row_start.0)
-                .map_or(false, |annots|annots.contains(
+                .is_some_and(|annots|annots.contains(
                     &LintAnnotation::Allow(error.rule_type)))
         }
     );
@@ -564,7 +587,7 @@ pub mod tests {
         crate::logging::init();
 
         let source =
-            "
+"
 dml 1.4;
 
 // dml-lint: allow-file=nsp_unary