+++ src/lint/rules/tests/mod.rs
@@ -4,3 +4,5 @@
 pub mod common;
 mod indentation;
 mod spacing;
+mod linelength;
+