+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 # Change Log
 
 # 0.9.15
-
+- Added support for line length and breaking rules regarding line-breaks after opening parentheses, method output arguments, conditional expressions and binary operands.
+- Added support for indendation rule indent_continuation_line.
 
 ## 0.9.14
 - Slight optimization to the memory usage of device-level analysis which